feat: support calling specific function from python provider
