+++ src/providers.ts
@@ -103,7 +103,7 @@ export async function loadApiProvider(
     const scriptPath = providerPath.split(':')[1];
     ret = new ScriptCompletionProvider(scriptPath, providerOptions);
   } else if (providerPath.startsWith('python:')) {
-    const scriptPath = providerPath.split(':')[1];
+    const scriptPath = providerPath.split(':').slice(1).join(':');
     ret = new PythonProvider(scriptPath, providerOptions);
   } else if (providerPath.startsWith('openai:')) {
     // Load OpenAI module
