+++ src/providers/pythonCompletion.ts
@@ -11,7 +11,7 @@ import type {
   ProviderEmbeddingResponse,
   ProviderClassificationResponse,
 } from '../types';
-import { safeJsonStringify, sha256 } from '../util';
+import { parsePathOrGlob, safeJsonStringify, sha256 } from '../util';
 
 interface PythonProviderConfig {
   pythonExecutable?: string;
@@ -19,12 +19,22 @@ interface PythonProviderConfig {
 
 export class PythonProvider implements ApiProvider {
   private config: PythonProviderConfig;
+  private scriptPath: string;
+  private functionName: string | null;
 
   constructor(
-    private scriptPath: string,
+    runPath: string,
     private options?: ProviderOptions,
   ) {
-    this.id = () => options?.id ?? `python:${this.scriptPath}`;
+    const { filePath: promptPath, functionName } = parsePathOrGlob(
+      options?.config.basePath || '',
+      runPath,
+    );
+    this.scriptPath = promptPath;
+    this.functionName = functionName || null;
+    this.id = () =>
+      options?.id ??
+      `python:${this.scriptPath}:${this.functionName ? `${this.functionName}` : 'default'}`;
     this.config = options?.config ?? {};
   }
 
@@ -66,39 +76,40 @@ export class PythonProvider implements ApiProvider {
       logger.debug(
         `Running python script ${absPath} with scriptPath ${this.scriptPath} and args: ${safeJsonStringify(args)}`,
       );
+      const functionName = this.functionName || apiType;
       let result;
       switch (apiType) {
         case 'call_api':
-          result = (await runPython(absPath, apiType, args, {
+          result = (await runPython(absPath, functionName, args, {
             pythonExecutable: this.config.pythonExecutable,
           })) as ProviderResponse;
           if (!result || (!('output' in result) && !('error' in result))) {
             throw new Error(
-              `The Python script \`${apiType}\` function must return a dict with an \`output\` string/object or \`error\` string, instead got: ${JSON.stringify(
+              `The Python script \`${functionName}\` function must return a dict with an \`output\` string/object or \`error\` string, instead got: ${JSON.stringify(
                 result,
               )}`,
             );
           }
           break;
         case 'call_embedding_api':
-          result = (await runPython(absPath, apiType, args, {
+          result = (await runPython(absPath, functionName, args, {
             pythonExecutable: this.config.pythonExecutable,
           })) as ProviderEmbeddingResponse;
           if (!result || (!('embedding' in result) && !('error' in result))) {
             throw new Error(
-              `The Python script \`${apiType}\` function must return a dict with an \`embedding\` array or \`error\` string, instead got ${JSON.stringify(
+              `The Python script \`${functionName}\` function must return a dict with an \`embedding\` array or \`error\` string, instead got ${JSON.stringify(
                 result,
               )}`,
             );
           }
           break;
         case 'call_classification_api':
-          result = (await runPython(absPath, apiType, args, {
+          result = (await runPython(absPath, functionName, args, {
             pythonExecutable: this.config.pythonExecutable,
           })) as ProviderClassificationResponse;
           if (!result || (!('classification' in result) && !('error' in result))) {
             throw new Error(
-              `The Python script \`${apiType}\` function must return a dict with a \`classification\` object or \`error\` string, instead of ${JSON.stringify(
+              `The Python script \`${functionName}\` function must return a dict with a \`classification\` object or \`error\` string, instead of ${JSON.stringify(
                 result,
               )}`,
             );
